feat-example-app-fix (#196)
### TL;DR
Added an Example Application to allow for testing and viewing of a recorder. This will allow users to see how to connect their own applications to the MeetingBot Service.

This is a fix from the `feat-example-app` branch, since there were some difficulties merging due to using both GitHub/Graphite being used, and test code which was replaced in other PRs.

Completes MEE-170

### What changed?
- New Next.JS app at `src/example-app`

### How to test?
1. Read instructions in the `README.md` file.


### Why make this change?
Ensuring that contributors have an example to base their own work off is vital for longevity of the project.